# Office-management-railway-deployment/swagger.yaml
openapi: 3.0.0
info:
  title: Auto Computation Office Management API
  version: 3.0.0
  description: API documentation for the Office Management System Backend
servers:
  - url: http://localhost:3000
    description: Local server
  - url: https://office-management-sbuh.onrender.com
    description: Production server (Render)
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: JWT token stored in a cookie
security:
  - cookieAuth: []
paths:
...
  /admin/emp/addEmp:
    post:
      summary: Add a new employee
      tags:
        - Admin-Employees
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - designation
                - phone
                - location
                - joining_date
                - salary
                - skills
                - employment_type
                - role
              properties:
                name:
                  type: string
                email:
                  type: string
                designation:
                  type: string
                phone:
                  type: string
                location:
                  type: string
                joining_date:
                  type: string
                  format: date
                salary:
                  type: number
                skills:
                  type: array
                  items:
                    type: string
                employment_type:
                  type: string
                department_id:
                  type: integer
                role:
                  type: string
      responses:
        "200":
          description: Employee added successfully
        "400":
          description: Validation error
        "500":
          description: Server error
  /admin/emp/updateEmp/{id}:
    put:
      summary: Update an existing employee
      tags:
        - Admin-Employees
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                designation:
                  type: string
                phone:
                  type: string
                location:
                  type: string
                joining_date:
                  type: string
                  format: date
                salary:
                  type: number
                skills:
                  type: array
                  items:
                    type: string
                employment_type:
                  type: string
      responses:
        "200":
          description: Employee updated successfully
        "404":
          description: Employee not found
        "500":
          description: Server error
  /admin/emp/all:
    get:
      summary: Get all active employees
      tags:
        - Admin-Employees
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of active employees
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      type: object
        "500":
          description: Server error
  /admin/emp/removeEmp/{id}:
    post:
      summary: Remove (archive) an employee
      tags:
        - Admin-Employees
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reason
                - password
              properties:
                reason:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Employee removed successfully
        "401":
          description: Unauthorized
        "500":
          description: Server error
  /admin/emp/allPastEmp:
    get:
      summary: Get all past (archived) employees
      tags:
        - Admin-Employees
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of past employees
        "500":
          description: Server error
  /admin/emp/permanentRemoveEmp/{id}:
    delete:
      summary: Permanently delete an employee from archive
      tags:
        - Admin-Employees
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee deleted permanently
        "500":
          description: Server error
...
  /admin/emp/all/meetings:
    get:
      summary: Get all employees for meeting scheduling (including departments)
      tags:
        - Admin-Employees
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of employees with department names
        "500":
          description: Server error
...